Version 0.0.64
- added VS2015 solution for windows 8.1 executables
- added experimental VS2017 solution
- 32 bit windows executables
- refactored types

NOTE: this is a revised release. Not sure why, but last time round the github desktop missing out some source code changes, and I didn't spot the omission.

+++ recipe/heater/webpage/fileindex.h
@@ -25,7 +25,8 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 // issues: not just / versus \  but also root version X: and \\xyzzy\ and how
 // WSL and other utilities mount windows volumes in 'nix
 
-typedef uint64_t fileindex_t;
+typedef size_t fileindex_t;
+// typedef uint64_t fileindex_t;
 typedef unsigned int fileindex_flags;
 class directory;
